BACKEND_1/requirements.txt
Flask
Pillow
pytesseract
pymongo
flask-cors
python-dotenv
Flask-Login
Werkzeug
Flask-JWT-Extended